Fixed Timer Issue

+++ sql/tables/Settings.sql
@@ -1,4 +1,4 @@
 CREATE TABLE Settings
 (ID INT,
-TimeLimit INT,
+TimeLimit FLOAT,
 IgnoreTimeLimit INT);
